+++ micro-ui/web/micro-ui-internals/packages/modules/core/src/pages/citizen/Login/index.js
@@ -96,7 +96,7 @@ const Login = ({ stateCode, isUserRegistered = true }) => {
     )
   );
 
-  const getUserType = () => Digit.UserService.getType();
+  const getUserType = () => "citizen" || Digit.UserService.getType();
 
   const handleOtpChange = (otp) => {
     setParmas({ ...params, otp });