+++ micro-ui/web/micro-ui-internals/packages/modules/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog 
 
+## [1.8.28]  [27-Feb-2025]
+- BUGFIX/SN-162 : Citizen sms otp fix
+
+## [1.8.28]  [27-Feb-2025]
+- BUGFIX/SN-162 : Citizen otp fix in the Login page.
+
+## [1.8.26]  [21-Feb-2025]
+- Publishing a new version for more stability & as part of Components Release
+
 ## [1.8.25]  [13-Feb-2025]
 - Added id prop for Privacy Component Checkbox